+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `path.pointInside(point)`
 - Added `path.bounds()`
 - Added `path.controlPointBounds()`
+- Added `path.reverse()`
 
 ## [0.2.0] - 2020-09-26