+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added `path.transform(transform, center=(0, 0))`
 - Added `path.arc(center, radius, startAngle, endAngle, clockwise)`
 - Added `path.arcTo(point1, point2, radius)`
+- Fixed bug when `font()` was not set
 
 ## [0.2.0] - 2020-09-26